(PC-29363)[PRO] fix: do not initialize venue form with acces libre data
