+++ pro/src/components/IndividualOfferForm/UsefulInformations/Venue/Venue.tsx
@@ -8,7 +8,7 @@ import {
 import FormLayout from 'components/FormLayout'
 import { IndividualOfferFormValues } from 'components/IndividualOfferForm'
 import { IndividualOfferForm } from 'components/IndividualOfferForm/types'
-import { buildAccessibilityFormValues } from 'pages/VenueEdition/setInitialFormValues'
+import { buildAccessibilityFormValues } from 'components/IndividualOfferForm/utils/setDefaultInitialFormValues'
 import { Select } from 'ui-kit'
 
 import { buildOffererOptions, buildVenueOptions } from './utils'
